requirements.txt
Flask>=2.3.0,<3.0.0
gunicorn>=21.0.0,<22.0.0
cryptography>=41.0.0,<42.0.0
bcrypt>=4.0.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
Werkzeug>=2.3.0,<3.0.0
Jinja2>=3.1.0,<4.0.0
MarkupSafe>=2.1.0,<4.0.0
itsdangerous>=2.1.0,<3.0.0
click>=8.1.0,<9.0.0
blinker>=1.6.0,<2.0.0
requests>=2.31.0,<3.0.0
